requirements.txt
altair<5
apprise==1.6.0
bottle==0.12.25
colorama==0.4.4
datetime
python-dateutil
gradio==3.41.0
inotify
librosa
matplotlib
numpy
paho-mqtt
pandas
pip
plotly
pytest-mock==3.7.0
pytest==7.1.2
pytz
requests
resampy
seaborn
setuptools
soundfile
streamlit==1.31.0
suntime
tflite-runtime
tzlocal
urllib3
wheel